addToQueue i addToWatched
1. Zmieniłam nazwy plików na camelCase
2. Przyciski dodawania działają za pierwszym kliknięciem
