+++ src/index.js
@@ -1,7 +1,7 @@
 import './js/fetchMovies.js';
 import './js/moviesContainer.js';
 import './js/footerStudents.js';
-import './js/add-to-queue.js'; // ymczasowy błąd 404
-import './js/watched.js';
+import './js/addToQueue.js'; // ymczasowy błąd 404
+import './js/addToWatched.js';
 import './js/fetchKeyWords';
 import './sass/main.scss';
